LayoutTests/platform/gtk/imported/w3c/web-platform-tests/inert/inert-retargeting-iframe.tentative-expected.txt

CONSOLE MESSAGE: Error: assert_equals: expected Element node <fieldset id="ancestorContainer">
    <legend>Non-inert a... but got Element node <button id="foreground" class="clicked">foreground</button>
CONSOLE MESSAGE: Error: assert_equals: expected Element node <fieldset id="ancestorContainer">
    <legend>Non-inert a... but got Element node <button id="foreground" class="clicked">foreground</button>
CONSOLE MESSAGE: Error: assert_equals: expected Element node <fieldset id="ancestorContainer">
    <legend>Non-inert a... but got Element node <button id="foreground" class="clicked">foreground</button>
CONSOLE MESSAGE: Error: assert_equals: expected Element node <fieldset id="ancestorContainer">
    <legend>Non-inert a... but got Element node <button id="foreground" class="clicked">foreground</button>
CONSOLE MESSAGE: Error: assert_equals: expected Element node <fieldset id="ancestorContainer">
    <legend>Non-inert a... but got Element node <button id="foreground" class="clicked">foreground</button>
CONSOLE MESSAGE: Error: assert_equals: expected Element node <fieldset id="ancestorContainer">
    <legend>Non-inert a... but got Element node <button id="foreground" class="clicked">foreground</button>
CONSOLE MESSAGE: Error: assert_equals: expected Element node <fieldset id="ancestorContainer">
    <legend>Non-inert a... but got Element node <button id="foreground" class="clicked">foreground</button>
CONSOLE MESSAGE: Error: assert_equals: expected Element node <fieldset id="ancestorContainer">
    <legend>Non-inert a... but got Element node <button id="foreground" class="clicked">foreground</button>
CONSOLE MESSAGE: Error: assert_equals: expected Element node <fieldset id="ancestorContainer">
    <legend>Non-inert a... but got Element node <button id="foreground" class="clicked">foreground</button>
CONSOLE MESSAGE: Error: assert_equals: expected Element node <fieldset id="ancestorContainer">
    <legend>Non-inert a... but got Element node <button id="foreground" class="clicked">foreground</button>
CONSOLE MESSAGE: Error: assert_equals: expected Element node <fieldset id="ancestorContainer">
    <legend>Non-inert a... but got Element node <button id="foreground" class="clicked">foreground</button>
CONSOLE MESSAGE: Error: assert_equals: expected Element node <fieldset id="ancestorContainer">
    <legend>Non-inert a... but got Element node <button id="foreground" class="clicked">foreground</button>
CONSOLE MESSAGE: Error: assert_equals: expected Element node <fieldset id="ancestorContainer">
    <legend>Non-inert a... but got Element node <button id="foreground" class="clicked">foreground</button>
Click on "Inert button".

The blue square ("Non-inert ancestor container") should turn green.
The yellow, semi-transparent square ("Non-inert, displaced container") should not turn red.
"Non-inert button" should not turn red.
(The full test suite checks a range of events.)

Non-inert ancestor container
 Non-inert, displaced ancestor
Inert container
foreground

Harness Error (FAIL), message = Error: assert_equals: expected Element node <fieldset id="ancestorContainer">
    <legend>Non-inert a... but got Element node <button id="foreground" class="clicked">foreground</button>

PASS elementsFromPoint returns inert element
FAIL Tests that any mouse event on a non-inert element is correctly targeted to that element promise_test: Unhandled rejection with value: "did not get auxclick on foreground"
FAIL Tests that any mouse event on an inert element is targeted to the nearest non-inert ancestor at the same coordinates promise_test: Unhandled rejection with value: "did not get mouseover on ancestorContainer"
PASS Tests that any pointer event on a non-inert element is correctly targeted to that element
TIMEOUT Tests that any pointer event on an inert element is targeted to the nearest non-inert ancestor at the same coordinates Test timed out

Harness Error (FAIL), message = Error: assert_equals: expected Element node <fieldset id="ancestorContainer">
    <legend>Non-inert a... but got Element node <button id="foreground" class="clicked">foreground</button>

PASS elementsFromPoint returns inert element
FAIL Tests that any mouse event on a non-inert element is correctly targeted to that element promise_test: Unhandled rejection with value: "did not get auxclick on foreground"
FAIL Tests that any mouse event on an inert element is targeted to the nearest non-inert ancestor at the same coordinates promise_test: Unhandled rejection with value: "did not get mouseover on ancestorContainer"
PASS Tests that any pointer event on a non-inert element is correctly targeted to that element
TIMEOUT Tests that any pointer event on an inert element is targeted to the nearest non-inert ancestor at the same coordinates Test timed out
